dataprep/pyproject.toml
[project]
name = "dataprep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "datasets>=3.6.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
]
